DBZ-8924 Move offset context loader to buffered package
